move context to its own file
